Revert "Adding import modal"
This reverts commit cd04e89a9778c37fcb4e926aeb7dafa44b3cd5dc.
