--- src/components/import-modal/import-modal.less
@@ -1,29 +1,0 @@
-@import (reference) "~less/compass/_theme.less";
-
-.import-modal {
-  &-output {
-    font-weight: bold;
-    font-size: 14px;
-  }
-
-  &-file {
-    font-weight: bold;
-    font-size: 14px;
-  }
-
-  &-type-selector {
-    button {
-      outline: none;
-      margin-right: 20px;
-      width: 150px;
-    }
-
-    button:focus {
-      outline: none;
-    }
-
-    .selected {
-      border: 2px solid #6BA442;
-    }
-  }
-}
