--- src/components/import-modal/index.js
@@ -1,2 +1,0 @@
-import ImportModal from './import-modal';
-export default ImportModal;
